add dashboard migrations server telemetry
